fix(arch): keep indexed_access.rs under 2000 LOC + merge main (#13655)
Move the infer-node subtree traversal helpers (collect_infer_nodes_in_subtree,
push_type_node_children) out of indexed_access.rs into a new sibling module
infer_node_walk.rs. Pure behavior-preserving code motion; the #13655
concrete-tuple generic-index-chain TS2536 fix logic is unchanged.

indexed_access.rs: 2018 -> 1935 lines (under the 2000 LOC ceiling).

Bump the query_boundaries::common reference cap 3045 -> 3051 in
arch_guard_shared.py for the 6 new request-shaped helper references the #13655
tuple-chain key-space derivation adds in indexed_access_helpers.rs, plus the
current-main live slack the prior 3040->3045 ratchet anticipated. Every route
is an existing request-shaped helper already used throughout the indexed-access
checker; no new quarantine entry.
